Remove a few compiler warnings.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@31587 72102866-910b-0410-8b05-ffd578937521

+++ Headers/AppKit/NSLayoutManager.h
@@ -97,17 +97,17 @@ typedef enum {
 - (void) textContainerChangedTextView: (NSTextContainer *)aContainer;
 
 
-- (NSPoint) locationForGlyphAtIndex: (unsigned int)glyphIndex;
+- (NSPoint) locationForGlyphAtIndex: (NSUInteger)glyphIndex;
 
 
 - (NSRect *) rectArrayForGlyphRange: (NSRange)glyphRange
 	withinSelectedGlyphRange: (NSRange)selGlyphRange
 	inTextContainer: (NSTextContainer *)container
-	rectCount: (unsigned int *)rectCount;
+	rectCount: (NSUInteger *)rectCount;
 - (NSRect *) rectArrayForCharacterRange: (NSRange)charRange
 	withinSelectedCharacterRange: (NSRange)selCharRange
 	inTextContainer: (NSTextContainer *)container
-	rectCount: (unsigned int *)rectCount;
+	rectCount: (NSUInteger *)rectCount;
 
 - (NSRect) boundingRectForGlyphRange: (NSRange)glyphRange 
 	inTextContainer: (NSTextContainer *)aTextContainer;
